Single column mmore articles working
more articles in the old system was using the article-artcle rel backwards, which is confusing.
Since it's article to article, technically, it could be used either way.  The naming implies one
way, but it was use the other.  The editors have added related articles to articles, and it was done
this backward way, so I have to maintain it in the new code.  Perhaps someday the names can change.
Really it's just a naming thing that leads to confusion.  oh well.