About page, initial work
Model and Admin built, some future work coming, but most is there.
Front end is wired up, but incomplete.

To continue on this work, the Person - Author extraction needs to happen next, then a Person m2m added ot About for the staff, like the partners
this will take the place of the pages/bio system in the mezz version.