fkchooser app to put a popup changelist on a FK field in the admin
In your models, define your FKs with PopupForeignKey
This will put a little image button next to the FK menu that when clicked will open the app and model's changelist
a Select button is added to the CL where the checkboxes used to be.  Clicking Select will set the value in the FK menu on the parent
page, and close the child dialog window.