follow button punt - redirects to /subscribe/
This history here is that maintaining a DB of subscribers locally, and on MC was tough.
The CALmatters staff changes the subscriber DB on MC, and users change it in both local and MC
The plan forward is to redesign it.  But for now, users get to choose categories in the MC
confirmation process.